+++ R/need.R
@@ -1,19 +1,38 @@
 
-#' Declare that a package is needed
+#' Declare that packages are needed
 #'
-#' Declare that one or more packages are required by subsequent functionality,
+#' Declare that one or more packages are required by subsequent functionality;
 #' and if they're missing, either prompt the user to install them, or exit with
 #' an informative error message.
 #'
-#' @inheritParams check_installed
-#' @param ... Names of required packages, as character strings.
+#' @inheritParams is_installed
+#' @param ... Names of required packages, as character strings. You can require
+#'   a minimum version by appending `>=[version]` to a package name - see
+#'   Examples.
+#' @param msg Custom message to display; if `NULL`, an informative one will be
+#'   constructed.
+#' @param install_cmd Installation command to run, as a call (i.e. probably
+#'   wrapped with [quote()] or [substitute()]). If `NULL`, [install.packages()]
+#'   will be used for package installation.
 #' @param ask Whether to give the user the option of installing the required
 #'   packages immediately.
 #'
-#' @return Invisibly, the names of any packages which were installed.
+#' @returns Invisibly, any package names from `...` which were installed.
 #'
-#' @examples \dontrun{
+#' @examples
+#' \dontrun{
+#'   need("dplyr")
+#'   need("dplyr", "tidyr")
 #'
+#'   # All unnamed arguments will be combined into one list of package names
+#'   shared_deps <- c("dplyr", "tidyr")
+#'   need(shared_deps, "stringr") # same as need("dplyr", "tidyr", "stringr")
+#'
+#'   # You can require a minimum version for some or all packages
+#'   need("dplyr>=1.0.0", "tidyr")
+#'
+#'
+#'   # Typically you'll want to use need() within a function
 #'   read_data <- function(path, clean_names = FALSE) {
 #'
 #'     # Call need() as early as possible, to avoid wasted work
@@ -27,45 +46,66 @@
 #'
 #'     output
 #'   }
-#'  }
+#'
+#'
+#'   # You can provide a custom message and/or installation command if needed
+#'   need(
+#'     "dplyr",
+#'     msg = "We need the development version of dplyr, for now!",
+#'     install_cmd = quote(remotes::install_github("tidyverse/dplyr"))
+#'   )
+#'
+#' }
 #'
 #' @export
+need <- function(
+    ...,
+    msg = NULL,
+    install_cmd = NULL,
+    ask = interactive(),
+    load = FALSE,
+    lib.loc = NULL
+) {
 
-need <- function(..., ask = interactive(), load = FALSE, lib.loc = .libPaths()) {
 
-  pkgs <- sort(unlist(c(...)))
+  pkgs <- sort(unique(unlist(c(...))))
 
-  installed <- check_installed(pkgs, load = load, lib.loc = lib.loc)
+  installed <- is_installed(pkgs, load = load, lib.loc = lib.loc)
 
   # Nothing to do if all packages are already present!
   if (all(installed))
     return(invisible(character()))
 
-  message(
-    "Additional packages are required to use this functionality: ",
-    paste0(pkgs[!installed], collapse = ", "),
-    "\n"
-  )
 
+  if (is.null(msg))
+    msg <- paste(
+      "To use this functionality, the following packages must be installed/updated:",
+      sprintf("\n\n  %s", paste0(pkgs[!installed], collapse = ", "))
+    )
+
+  message(msg, "\n")
 
   if (isTRUE(ask))
     choice <- utils::menu(
-      title = "Would you like to install these packages now?",
+      title = "Would you like to install/update now?",
       choices = c("Yes", "No")
     )
 
 
+  # Default installation command, if a bespoke one wasn't provided
+  if (is.null(install_cmd))
+    # Not explicitly stating utils::install.packages() because we might have a
+    # shim available e.g. from {renv}
+    install_cmd <- substitute(install.packages(p), list(p = names(installed)[!installed]))
+
 
   if (!isTRUE(ask) || choice != 1)
     stop(
-      "Please install the following packages to use this functionality: ",
-      "\n\n  install.packages(c(",
-      paste0(dQuote(pkgs[!installed], q = FALSE), collapse = ", "),
-      "))",
+      "Please install/update packages to use this functionality:",
+      sprintf("\n\n  %s", deparse(install_cmd)),
       call. = FALSE
     )
 
-
-  utils::install.packages(pkgs[!installed])
+  eval(install_cmd)
   invisible(pkgs[!installed])
 }